# Speed up junk detection