Before this patch visitors traversed whole translation unit for each mutant,
within this patch, Mull traverses the only function that can potentially
contain the mutant.

---
On an example app I've got the following numbers (debug build):
Before:
 - junk detection: 72s
 - total: 108s
After:
 - junk detection: 4s
 - total: 40s


@stanislaw could you give it a try on your project and check is there is no regression?
I would also appreciate if you can share the numbers.