@AlexDenisov 2019-10-04
Given the number of mutants you have this should be a good manual test: could you make sure that the number of mutants after junk detection is the same on the master branch and on this branch?