@stanislaw 2019-10-04
So, the old vs new numbers:

```
Running original tests (threads: 8): 71/71. Finished in 3157ms.
Searching mutants across functions (threads: 8): 5923/5923. Finished in 69ms.
Applying filter: no debug info (threads: 8): 42737/42737. Finished in 13ms.
Applying filter: file path (threads: 8): 36099/36099. Finished in 384ms.
Applying filter: junk (threads: 8): 28140/28140. Finished in 93285ms.
```

```
Running original tests (threads: 8): 71/71. Finished in 3151ms.
Searching mutants across functions (threads: 8): 5923/5923. Finished in 83ms.
Applying filter: no debug info (threads: 8): 42737/42737. Finished in 13ms.
Applying filter: file path (threads: 8): 36099/36099. Finished in 401ms.
Applying filter: junk (threads: 8): 28140/28140. Finished in 30963ms.
```