@stanislaw 2019-10-04
It looks like the number of mutations have changed:

```
...Searching mutants across functions (threads: 8): 5923/5923. Finished in 69ms.
Applying filter: no debug info (threads: 8): 42737/42737. Finished in 13ms.
Applying filter: file path (threads: 8): 36099/36099. Finished in 384ms.
Applying filter: junk (threads: 8): 28140/28140. Finished in 93285ms.
Preparing mutations (threads: 1): 1/1. Finished in 227280ms.
Applying mutations (threads: 1): 468/468. Finished in 33ms.
Compiling original code (threads: 8): 49/49. Finished in 131511ms.
Running mutants (threads: 8): 158/468
```

```
Applying filter: file path (threads: 8): 36099/36099. Finished in 361ms.
Applying filter: junk (threads: 8): 28140/28140. Finished in 30534ms.
Preparing mutations (threads: 1): 1/1. Finished in 215727ms.
Applying mutations (threads: 1): 353/353. Finished in 32ms.
Compiling original code (threads: 8): 49/49. Finished in 55ms.
Running mutants (threads: 8): 353/353. Finished in 150692ms.
Mutation Testing Elements reporter: generating report to /tmp/strict-doc.json
```