@AlexDenisov 2019-10-16
I think we actually have a buggy implementation on the master branch, so this PR is the correct one.
@stanislaw did you do any further debugging?