@stanislaw 2019-10-16
>  @stanislaw did you do any further debugging?

No, I have not done a good analysis since then unfortuntately. The last step was to pass the master vs this PR JSON outputs to you.