@stanislaw 2019-10-27
I am using the following Python script to check for the differences between two Elements reports:

The main difference that I see between two output elements reports as I am running them on my pet project is that too many of the mutations are duplicated this is why it is hard to tell which branch: master or this one works correct.
 
```
#!/usr/local/bin/python3

import json
import os
import sys

from pprint import pprint

def hash_4(a, b, c, d):
    assert(a > 0)
    assert(b > 0)
    assert(c > 0)
    assert(d > 0)
    return "{}:{},{}:{}".format(a, b, c, d)

def build_map(json):
    mutations_map = {}

    files_json = json['files']

    for file, file_content in files_json.items():
        mutations_map[file] = {}

        file_mutants = []
        for idx, mutant in enumerate(file_content['mutants']):
            # print(mutant)

            start_line = mutant['location']['start']['line']
            start_column = mutant['location']['start']['column']
            end_line = mutant['location']['end']['line']
            end_column = mutant['location']['end']['column']
            status = mutant['status']
            assert status

            hash = hash_4(start_line, start_column, end_line, end_column)
            # print(hash)
            if not status in mutations_map[file]:
                mutations_map[file][status] = {}

            if not hash in mutations_map[file][status]:
                mutations_map[file][status][hash] = idx

    return mutations_map

def remove_common(json, mutations_map):
    files_json = json['files']
    for file, file_content in files_json.items():
        mutants = file_content['mutants']

        file_map = mutations_map[file] if file in mutations_map else {}

        for idx, mutant in reversed(list(enumerate(mutants))):
            start_line = mutant['location']['start']['line']
            start_column = mutant['location']['start']['column']
            end_line = mutant['location']['end']['line']
            end_column = mutant['location']['end']['column']
            status = mutant['status']

            assert start_line
            assert start_column
            assert status

            hash = hash_4(start_line, start_column, end_line, end_column)

            if status in file_map:
                if hash in file_map[status]:
                    # pass
                    # del(file_map[hash])
                    del(file_map[status][hash])
                    del(mutants[idx])
                else:
                    pass # assert None

def dump_map(map):
    unique = {}
    for file, file_map in map.items():
        print(file)
        for status, mutations in file_map.items():
            for hash, _ in mutations.items():
                print(hash)
                print(mutations)
                # assert not hash in unique
                unique[hash] = 1

def main():
    if len(sys.argv) != 3:
        print("two arguments: in and out")
        exit(1)

    lhs_file = sys.argv[1]
    rhs_file = sys.argv[2]

    if not os.path.isfile(lhs_file):
        print("lhs file does not exist: {}".format(lhs_file))

    if not os.path.isfile(rhs_file):
        print("rhs file does not exist: {}".format(lhs_file))

    with open(lhs_file) as f:
        lhs_json = json.load(f)

    with open(rhs_file) as f:
        rhs_json = json.load(f)

    lhs_map = build_map(lhs_json)
    rhs_map = build_map(rhs_json)

    dump_map((lhs_map))
    dump_map((rhs_map))

    remove_common(lhs_json, rhs_map)
    remove_common(rhs_json, lhs_map)

    with open('out-lhs.json', 'w') as f:
        json.dump(lhs_json, f)

    with open('out-rhs.json', 'w') as f:
        json.dump(rhs_json, f)

    # print(lhs_files_json)

main()
```