@AlexDenisov 2019-11-06
After more debugging, it seems like the discrepancy is coming from 'broken' mutators (such as scalar value).
I'll get it in and then we can have another session whenever AST filter is in place.